+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 
+- Drop support for Python 2.7. (**breaking change**)
 - Don't mutate ``models`` argument when passed in as a list to ``SQLClient.save|core.save``.
 - Allow generators to be passed into ``SQLClient.save|core.save`` and ``SQLClient.destroy|core.destroy``.
 - Remove deprecated methods: (**breaking change**)